requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bert-score==0.3.13
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
datasets==2.15.0
dill==0.3.7
distro==1.8.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faiss-cpu==1.7.3
fasttext==0.9.2
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
futures==3.0.5
gitdb==4.0.11
GitPython==3.1.40
goslate==1.5.4
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-resources==6.1.1
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mauve-text==0.3.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.6.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.7
pyarrow==14.0.2
pyarrow-hotfix==0.6
pybind11==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
PyDictionary==2.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
sentencepiece==0.1.99
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
string2string==0.0.150
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
